+++ gofdesignpattern/src/test/kotlin/com/example/gofdesignpattern/week04_builder/TossBankTransfer.kt
@@ -3,7 +3,7 @@ package com.example.gofdesignpattern.week04_builder
 class TossBankTransfer(
     override val fromBankCode: Int,
     override val toBankCode: Int,
-    override val amount: Int,
+    override val amount: Long,
     override val summary: String?,
     override val name: String
 ) : Transfer {
